requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1